[commons] clean code

+++ vertigo-commons-impl/src/test/java/io/vertigo/commons/cache/CacheManagerInitializer.java
@@ -41,8 +41,8 @@ public final class CacheManagerInitializer implements ComponentInitializer {
 		//Paramétrage d'un cache spécifique au test
 		/** Parametre du cache, pour une config ou il est multi-session*/
 		final int maxElementsInMemory = 50000;
-		final long timeToLiveSeconds = 1000; //longévité d'un élément
-		final long timeToIdleSeconds = 10; //longévité d'un élément non utilisé
+		final int timeToLiveSeconds = 1000; //longévité d'un élément
+		final int timeToIdleSeconds = 10; //longévité d'un élément non utilisé
 
 		manager.addCache(CONTEXT_EDITABLE, new CacheConfig("test", true, maxElementsInMemory, timeToLiveSeconds, timeToIdleSeconds));
 		manager.addCache(CONTEXT_READONLY, new CacheConfig("test", false, maxElementsInMemory, timeToLiveSeconds, timeToIdleSeconds));
